refactor(wwa-message): マクロ・ウィンドウの機能を別ファイルに切り出した (再)
オリジナルは c918214 です。コンフリクトを追従してやり直しました。
