+++ packages/engine/src/wwa_data.ts
@@ -1,6 +1,6 @@
+import { type WWAData } from "@wwawing/common-interface";
 import { WWA } from "./wwa_main";
 import { Camera } from "./wwa_camera";
-import { type WWAData } from "@wwawing/common-interface";
 import type { JsonResponseErrorKind } from "./json_api_client";
 
 export { type WWAData };
